store/tikv:move mockCommitErrorEnable from kv to tikv since it is used in tikv
