+++ kv/txn.go
@@ -17,7 +17,6 @@ import (
 	"context"
 	"math"
 	"math/rand"
-	"sync/atomic"
 	"time"
 
 	"github.com/pingcap/parser/terror"
@@ -92,24 +91,6 @@ func BackOff(attempts uint) int {
 	sleep := time.Duration(rand.Intn(upper)) * time.Millisecond
 	time.Sleep(sleep)
 	return int(sleep)
-}
-
-// mockCommitErrorEnable uses to enable `mockCommitError` and only mock error once.
-var mockCommitErrorEnable = int64(0)
-
-// MockCommitErrorEnable exports for gofail testing.
-func MockCommitErrorEnable() {
-	atomic.StoreInt64(&mockCommitErrorEnable, 1)
-}
-
-// MockCommitErrorDisable exports for gofail testing.
-func MockCommitErrorDisable() {
-	atomic.StoreInt64(&mockCommitErrorEnable, 0)
-}
-
-// IsMockCommitErrorEnable exports for gofail testing.
-func IsMockCommitErrorEnable() bool {
-	return atomic.LoadInt64(&mockCommitErrorEnable) == 1
 }
 
 // TxnInfo is used to keep track the info of a committed transaction (mainly for diagnosis and testing)
