+++ store/tikv/test_util.go
@@ -14,6 +14,8 @@
 package tikv
 
 import (
+	"sync/atomic"
+
 	"github.com/google/uuid"
 	"github.com/pingcap/errors"
 	pd "github.com/tikv/pd/client"
@@ -41,4 +43,22 @@ func NewTestTiKVStore(client Client, pdClient pd.Client, clientHijack func(Clien
 
 	tikvStore.mock = true
 	return tikvStore, errors.Trace(err)
+}
+
+// mockCommitErrorEnable uses to enable `mockCommitError` and only mock error once.
+var mockCommitErrorEnable = int64(0)
+
+// MockCommitErrorEnable exports for gofail testing.
+func MockCommitErrorEnable() {
+	atomic.StoreInt64(&mockCommitErrorEnable, 1)
+}
+
+// MockCommitErrorDisable exports for gofail testing.
+func MockCommitErrorDisable() {
+	atomic.StoreInt64(&mockCommitErrorEnable, 0)
+}
+
+// IsMockCommitErrorEnable exports for gofail testing.
+func IsMockCommitErrorEnable() bool {
+	return atomic.LoadInt64(&mockCommitErrorEnable) == 1
 }
